+++ app/src/backend/core/system/settings.service.js
@@ -64,6 +64,7 @@ settingDefinitions = {
   checkUpdatesInBackground: { type: "boolean", defaultValue: true },
   checkAppUpdatesOnStartup: { type: "boolean", defaultValue: true },
   wineCommand: { type: "string", defaultValue: null, nullable: true },
+  engineVersionPreferences: { type: "string", defaultValue: "{}" },
   firstRunStorageSetupComplete: { type: "boolean", defaultValue: false },
   baseGameSupportWarningShown: { type: "boolean", defaultValue: false },
 };